+++ test_py
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+export ROS_DOMAIN_ID=5
+for i in 1 2 4 8
+do
+	for j in 4 16 64 256 1024 4096 16384 65356
+	do
+		ros2 launch py_echo launch$i.py message_size:=$j publisher_max:=$i
+	done
+done
+
